+++ src/pytask_parallel/plugin.py
@@ -1,6 +1,7 @@
 """Entry-point for the plugin."""
 from __future__ import annotations
 
+from pluggy import PluginManager
 from pytask import hookimpl
 from pytask_parallel import build
 from pytask_parallel import config
@@ -9,7 +10,7 @@ from pytask_parallel import logging
 
 
 @hookimpl
-def pytask_add_hooks(pm):
+def pytask_add_hooks(pm: PluginManager) -> None:
     """Register plugins."""
     pm.register(build)
     pm.register(config)